requirements.txt
# Core dependencies
python-dotenv==1.0.0

# CLI framework
click==8.1.7
rich==13.7.0

# OpenAI API
openai==2.6.1

# API server
fastapi>=0.110.0
uvicorn[standard]>=0.25.0

# macOS APIs for screen capture, window metadata, and Vision OCR
pyobjc-framework-Quartz==12.0  # Screen capture
pyobjc-framework-Cocoa==12.0   # macOS APIs
pyobjc-framework-Vision==12.0  # Local OCR (Apple Vision framework)

# Database
sqlite-utils==3.36

# Vector embeddings, search, reranking
sentence-transformers>=5.0.0
transformers>=4.57.0,<5.0.0
chromadb>=0.5.4
numpy>=1.26.0,<2.0.0
FlagEmbedding>=1.2.10

# Async and concurrency
aiofiles==23.2.1

# Utilities
python-dateutil==2.8.2
psutil==5.9.6  # System monitoring

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Logging
structlog==23.2.0

# MCP server and Context7 (AI assistant integration + docs)
mcp>=1.19.0
httpx>=0.28.1
requests>=2.32.5

# Image processing (frame diffing + DeepSeek batch OCR)
Pillow==12.0.0
imagehash==4.3.2

# Local MLX VLM backend for DeepSeek OCR (Apple Silicon) - YOUR NEW FEATURE
# DeepSeek OCR = 10-20x cost reduction vs cloud
mlx-vlm>=0.3.5

# Video encoding (H.264/HEVC hardware-accelerated compression)
pyobjc-framework-AVFoundation==12.0
pyobjc-framework-CoreMedia==12.0
pyobjc-framework-VideoToolbox==12.0  # For HEVC/H.265 hardware encoding

# UI (Streamlit for real-time monitoring)
streamlit==1.40.2
